Moved over files from previous project
Created a WebGL build that works. Moved over the scripts from the previous unity project to this one, cleaning the code along the way. Created the test_index file to test the webcomponent and the wc-body-ui file. the WC will display the empty unity build in the window but only when the setUnityInstance and attributeChangeCallback are commented out