New build of the WebGL, created main.js and style.css, updated index
New build of the WebGL contains a cube with a script that rotates and pans on button click. Index has had gismos prepared to interact with the WebGL. The main.js is getting prepared to hold data passing functions when index interacted. When the style sheet is finished it will control the orientation of the index.